requirements/base.txt
Django==1.10.4
django-jet==1.0.4
django-mail-templated==2.6.5
djangorestframework==3.5.3
drfdocs==0.0.11
feedparser==5.2.1
Markdown==2.6.7
Pillow==3.4.2
psycopg2==2.6.2
Unipath==1.1